requirements.txt
fastapi
strawberry-graphql
uvicorn
sqlalchemy
psycopg2-binary